requirements3.11.txt
pygame==2.0.1
PyOpenGL==3.1.5
Pillow==8.3.2